Refactor message body storage to improve organization and clarity
This change organizes the message body storage implementation by:

*   Moving related files into a dedicated `Implementation.BodyStorage` namespace.
*   Extracting default container and key prefix names into `const` fields.
*   Adding XML documentation to key body storage components.
