+++ src/ServiceControl.Persistence.EFCore/Implementation/BodyStorage/S3BodyStorageInstaller.cs
@@ -1,4 +1,4 @@
-namespace ServiceControl.Persistence.EFCore.Implementation;
+namespace ServiceControl.Persistence.EFCore.Implementation.BodyStorage;
 
 using Amazon.S3.Model;
 using Amazon.S3.Util;
