+++ src/ServiceControl.Persistence.EFCore/Infrastructure/IBodyStoragePersistence.cs
@@ -4,8 +4,13 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 
-// Bodies are immutable and addressed by bodyId (the UniqueMessageId) alone, so re-failures of the
-// same message resolve to the same stored body and writes can be skipped when it already exists.
+/// <summary>
+/// External storage for message bodies too large or too binary to sit inline in the database.
+/// </summary>
+/// <remarks>
+/// Bodies are immutable and addressed by bodyId (the UniqueMessageId) alone, so re-failures of the
+/// same message resolve to the same stored body and writes can be skipped when it already exists.
+/// </remarks>
 public interface IBodyStoragePersistence
 {
     Task WriteBody(string bodyId, ReadOnlyMemory<byte> body, string contentType, CancellationToken cancellationToken = default);
